+++ CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0] - 2024-07-26
+
+### Added
+
+- Databricks pools to create must be specified.
+
+### Updated
+
+- Databricks pool modules.
+
 ## [4.3.0] - 2024-06-06
 
 ### Added